Moved claimant info into multipage form
