+++ examples/multi-page/index.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import { Route } from 'react-router-dom'
 import { FormRouter } from '@department-of-veterans-affairs/va-forms-system-core';
 import FormIntroductionPage from './intro-page';
-import PersonalInformationPage from './page-one';
+import ClaimantInformationPage from './claimant-information';
 import ContactInformationPage from './page-two';
 
 const NoMatch = (props) => (
@@ -18,7 +18,7 @@ const FormApp = (props) => {
     <div className='vads-u-display--flex vads-u-align-items--center vads-u-flex-direction--column'>
       <FormRouter basename={props.basename} formData={props.initialValues} title="Burials Example">
         <Route index element={<FormIntroductionPage />} />
-        <Route path="/page-one" element={<PersonalInformationPage />} />
+        <Route path="/page-one" element={<ClaimantInformationPage />} />
         <Route path="/page-two" element={<ContactInformationPage />} />
         <Route path="*" element={<NoMatch name="No Routes for App" />} />
       </FormRouter>
